refactoring project once more to better align with best-practice go structure
