+++ .github/workflows/status-checks.yml
@@ -53,7 +53,7 @@ jobs:
 
       - name: Run Tests
         if: matrix.task == 'test'
-        run: go test -v $(go list ./... | grep -v "api_calls.go,models/,repository/,transaction_service.go") 
+        run: go test -v $(go list ./... | grep -v "/cmd,/internal/app,/internal/config,/transactions") 
         # db & transaction_service tests will be written after full implementations
         
       - name: Install & Run Linter
