+++ internal/app/app.go
@@ -75,7 +75,7 @@ func loadConfig() {
 		env = "git_testing"
 	}
 
-	configFile := fmt.Sprintf("config/%s_config.yml", env)
+	configFile := fmt.Sprintf("../internal/config/%s_config.yml", env)
 
 	log.Printf("Loading config: %s", configFile)
 	data, err := os.ReadFile(configFile)
