--- internal/repository/transaction_repository.go
@@ -1,19 +1,0 @@
-package repository
-
-import (
-	// "log"
-
-	"github.com/undersleep7x/cryo-project/internal/models"
-)
-
-func SaveTransaction(txn models.Transaction) error {
-	return nil
-}
-
-func FindTransactionById(txnId string) models.Invoice {
-	return models.Invoice{}
-}
-
-func UpdateTransactionById(txx models.Transaction) error {
-	return nil
-}
