+++ internal/transactions/repository.go
@@ -1,0 +1,19 @@
+package transactions
+
+import (
+	// "log"
+
+
+)
+
+func SaveTransaction(txn Transaction) error {
+	return nil
+}
+
+func FindTransactionById(txnId string) Invoice {
+	return Invoice{}
+}
+
+func UpdateTransactionById(txx Transaction) error {
+	return nil
+}
